+++ lib/home/view/ledger/tab_bar_view.dart
@@ -49,7 +49,7 @@ class _LedgerTabBarViewState extends State<LedgerTabBarView>
         padding: const EdgeInsets.all(15.0),
         child: TabBarView(
           controller: _tabController,
-          children: [LedgerPage(), MarketPrice()],
+          children: [LedgerPage(), MarketPriceTable()],
         ),
       ),
     );
